Update README.md
Our project explains about Billing Management System. The main objective of the Java Project on Billing System is to manage the details of Items. The Project is based on the concept of calculating the total bill & the end generating the invoice of the customer. Talking about the project, it contains less but important features. A user can select any category, CRUD operations before calculating the bill. Our project includes some Modules: • Bill Module • Product Module We have developed this application in Java. It’s a pure Console based application. The main feature of the project is to make Billing easier and manage items in easier way.